# docs/openapi.yaml
openapi: 3.1.0
info:
  title: PROOF API
  description: PROOF 서비스 REST API 명세서
  version: 1.0.0
  contact:
    name: PROOF Team
servers:
  - url: http://localhost:8000
    description: 로컬 개발 서버
...
paths:
...
  /api/users/me:
    get:
      tags:
        - Users
      summary: 내 정보 조회
      description: 현재 로그인한 사용자의 정보를 조회합니다
      parameters:
        - $ref: '#/components/parameters/UserIdHeader'
      responses:
        '200':
          description: 사용자 정보
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'

    patch:
      tags:
        - Users
      summary: 내 정보 수정
      description: 사용자 정보를 수정합니다
      parameters:
        - $ref: '#/components/parameters/UserIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: "홍길동"
                university:
                  type: string
                  example: "서울대학교"
                major:
                  type: string
                  example: "경영학과"
      responses:
        '200':
          description: 수정 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'

  # ==================== Spaces ====================
  /api/spaces:
    get:
      tags:
        - Spaces
      summary: 스페이스 목록 조회
      description: 사용자의 스페이스 목록을 조회합니다
      parameters:
        - $ref: '#/components/parameters/UserIdHeader'
        - name: status
          in: query
          description: 상태 필터
          schema:
            type: string
            enum: [active, completed, archived]
      responses:
        '200':
          description: 스페이스 목록
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Space'

    post:
      tags:
        - Spaces
      summary: 스페이스 생성
      description: 새로운 스페이스를 생성합니다
      parameters:
        - $ref: '#/components/parameters/UserIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - type
                - start_date
                - end_date
                - reflection_cycle
              properties:
                name:
                  type: string
                  description: 스페이스 이름
                  example: "2024 마케팅 공모전"
                type:
                  type: string
                  enum: [공모전, 프로젝트, 동아리, 인턴십]
                  description: 활동 유형
                  example: "공모전"
                description:
                  type: string
                  description: 설명
                  example: "전국 대학생 마케팅 공모전"
                start_date:
                  type: string
                  format: date
                  description: 시작일
                  example: "2024-01-01"
                end_date:
                  type: string
                  format: date
                  description: 종료일
                  example: "2024-03-31"
                reflection_cycle:
                  type: string
                  enum: [매일, 주1회, 2주1회, 월1회]
                  description: 회고 주기
                  example: "주1회"
                reminder_enabled:
                  type: boolean
                  description: 리마인더 활성화
                  example: true
                keywords:
                  type: array
                  items:
                    type: string
                  description: 키워드
                  example: ["마케팅", "기획", "브랜딩"]
      responses:
        '201':
          description: 생성 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Space'
  /api/spaces/{space_id}:
    get:
      tags:
        - Spaces
      summary: 스페이스 상세 조회
      description: 특정 스페이스의 상세 정보를 조회합니다
      parameters:
        - $ref: '#/components/parameters/UserIdHeader'
        - name: space_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: 스페이스 ID
      responses:
        '200':
          description: 스페이스 상세 정보
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Space'

    patch:
      tags:
        - Spaces
      summary: 스페이스 수정
      description: 스페이스 정보를 수정합니다
      parameters:
        - $ref: '#/components/parameters/UserIdHeader'
        - name: space_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                status:
                  type: string
                  enum: [active, completed, archived]
                reflection_cycle:
                  type: string
                  enum: [매일, 주1회, 2주1회, 월1회]
                reminder_enabled:
                  type: boolean
      responses:
        '200':
          description: 수정 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Space'

    delete:
      tags:
        - Spaces
      summary: 스페이스 삭제
      description: 스페이스를 삭제합니다
      parameters:
        - $ref: '#/components/parameters/UserIdHeader'
        - name: space_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: 삭제 성공
  /api/spaces/{space_id}/members:
    get:
      tags:
        - Spaces
      summary: 스페이스 멤버 목록 조회
      description: 스페이스의 멤버 목록을 조회합니다
      parameters:
        - $ref: '#/components/parameters/UserIdHeader'
        - name: space_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 멤버 목록
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SpaceMember'

    post:
      tags:
        - Spaces
      summary: 스페이스 멤버 초대
      description: 스페이스에 멤버를 초대합니다
      parameters:
        - $ref: '#/components/parameters/UserIdHeader'
        - name: space_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - user_id
              properties:
                user_id:
                  type: string
                  description: 초대할 사용자 ID
                  example: "jane_doe"
                role:
                  type: string
                  enum: [owner, member]
                  default: member
                  example: "member"
      responses:
        '201':
          description: 초대 성공

  # ==================== Reflections ====================
...
  /api/templates:
    get:
      tags:
        - Templates
      summary: 템플릿 목록 조회
      description: 회고 템플릿 목록을 조회합니다
      parameters:
        - name: category
          in: query
          description: 카테고리 필터
          schema:
            type: string
            enum: [전체, 기본, 심화, 감정, 분석, 정기]
      responses:
        '200':
          description: 템플릿 목록
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReflectionTemplate'
  /api/templates/{template_id}:
    get:
      tags:
        - Templates
      summary: 템플릿 상세 조회
      description: 특정 템플릿의 상세 정보를 조회합니다
      parameters:
        - name: template_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 템플릿 상세 정보
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReflectionTemplate'
  /api/templates/recommend:
    get:
      tags:
        - Templates
      summary: 추천 템플릿 조회
      description: 사용자에게 적합한 템플릿을 추천합니다
      parameters:
        - $ref: '#/components/parameters/UserIdHeader'
      responses:
        '200':
          description: 추천 템플릿
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReflectionTemplate'

  # ==================== Activities ====================
...
components:
  parameters:
    UserIdHeader:
      name: x-user-id
      in: header
      required: true
      description: 사용자 ID
      schema:
        type: string
        example: "john_doe"
  schemas:
    User:
      type: object
      properties:
        user_id:
          type: string
          example: "john_doe"
        name:
          type: string
          example: "홍길동"
        email:
          type: string
          example: "john@example.com"
        university:
          type: string
          example: "서울대학교"
        major:
          type: string
          example: "경영학과"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Space:
      type: object
      properties:
        space_id:
          type: string
          format: uuid
        user_id:
          type: string
        name:
          type: string
          example: "2024 마케팅 공모전"
        type:
          type: string
          enum: [공모전, 프로젝트, 동아리, 인턴십]
        description:
          type: string
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
        reflection_cycle:
          type: string
          enum: [매일, 주1회, 2주1회, 월1회]
        reminder_enabled:
          type: boolean
        status:
          type: string
          enum: [active, completed, archived]
        keywords:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    SpaceMember:
      type: object
      properties:
        member_id:
          type: string
          format: uuid
        space_id:
          type: string
          format: uuid
        user_id:
          type: string
        role:
          type: string
          enum: [owner, member]
        joined_at:
          type: string
          format: date-time
...
    ReflectionTemplate:
      type: object
      properties:
        template_id:
          type: string
          format: uuid
        name:
          type: string
          example: "KPT 회고"
        category:
          type: string
          enum: [기본, 심화, 감정, 분석, 정기]
        description:
          type: string
        icon:
          type: string
          example: "target"
        questions:
          type: array
          items:
            type: object
            properties:
              question:
                type: string
              placeholder:
                type: string
        recommended_situations:
          type: array
          items:
            type: string
        usage_count:
          type: integer
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
...